fix(gateway): route local L2 timers to scene extraction (#227)
* fix(gateway): route local L2 timers to scene extraction


* Delete src/gateway/timer-routing.test.ts

---------
